FEATURE: Add custom label for auto created child node
This change add a node type configuration to generate a custom node label
for auto created child nodes:

    'TYPO3.Neos.NodeTypes:Page':
      superTypes:
        'TYPO3.Neos:Document': TRUE
      childNodes:
        main:
          label: "${'Main Content Collection'}"
          type: 'TYPO3.Neos:ContentCollection'

This allow to have a less technical content tree. This change include the
schema validation and default label for default node types.

The label is an EEL expression, like the normal node label. Inside the
expression you can access the parent node in the variable ```parentNode```.